Renamed model_export to model_packaging
