+++ model_server_tools/model_packaging/model_packaging.py
@@ -14,7 +14,7 @@ Command line interface to export model files to be used for inference by MXNet M
 
 import logging
 from .arg_parser import ArgParser
-from .export_model_utils import ModelExportUtils
+from .model_packaging_utils import ModelExportUtils
 
 
 def export_model(args, manifest, export_file_path=None):
