+++ model_server_tools/model_packaging/tests/unit_tests/test_model_packaging_utils.py
@@ -14,7 +14,7 @@ from mock import Mock, mock_open, patch
 import os
 import sys
 from collections import namedtuple
-from model_server_tools.model_packaging.export_model_utils import ModelExportUtils
+from model_server_tools.model_packaging.model_packaging_utils import ModelExportUtils
 from model_server_tools.model_packaging.manifest_components.engine import EngineType
 from model_server_tools.model_packaging.manifest_components.manifest import RuntimeType
 
@@ -70,7 +70,7 @@ class TestExportModelUtils:
         @pytest.fixture()
         def patches(self, mocker):
             Patches = namedtuple('Patches', ['utils', 'listdir'])
-            patch = Patches(mocker.patch('model_server_tools.model_packaging.export_model_utils.ModelExportUtils'),
+            patch = Patches(mocker.patch('model_server_tools.model_packaging.model_packaging_utils.ModelExportUtils'),
                             mocker.patch('os.listdir'))
 
             patch.listdir.return_value = set(['a', 'b', 'c'])
